将publish迁移到Video
