+++ TIKTOK_User/configs/config.go
@@ -1,9 +1,1 @@
 package configs
-
-/*
-minio info
-*/
-
-const minioIp = "http://120.25.2.146:9000/"
-const AccessKeyId = "minioadmin"
-const SecretAccessKey = "minioadmin"
