+++ TIKTOK_User/dal/mysql/publishDao.go
@@ -2,7 +2,6 @@ package mysql
 
 import (
 	"GoProject/model"
-	"time"
 )
 
 /*
@@ -14,17 +13,6 @@ func CreatePublishVideo(UserId, VideoId int64) error {
 	video := model.Publish{UserId: UserId, VideoId: VideoId}
 	result := DB.Create(&video)
 	return result.Error
-}
-
-/*
-CreateVideo
-根据UserId 和 VideoId url，在vms中添加一行数据
-*/
-func CreateVideo(UserId int64, playUrl string, coverUrl string, title string) (int64, error) {
-	// TODO : impl
-	video := model.Video{UserId: UserId, PlayUrl: playUrl, CoverUrl: coverUrl, Title: title, PublishTime: time.Now().Unix()}
-	result := DBV.Create(&video)
-	return video.Id, result.Error
 }
 
 /*
