+++ TIKTOK_User/service/publishService.go
@@ -3,7 +3,5 @@ package service
 import "GoProject/model/vo"
 
 type PublishService interface {
-	PublishVideo(userId int64, videoData []byte, videoTitle string) error
-
 	GetVideoList(userIdTar, userIdSrc int64) ([]vo.VideoInfo, error)
 }
