+++ TIKTOK_Video/configs/config.go
@@ -2,3 +2,8 @@ package configs
 
 // VideoCount 一次刷新获取的视频数量
 const VideoCount = 30
+
+// info for minio server
+
+const AccessKeyId = "minioadmin"
+const SecretAccessKey = "minioadmin"
